update chat helpers
